fixed bugs

+++ src/main/java/tn/esprit/spring/kaddem/services/ContratServiceImpl.java
@@ -87,7 +87,7 @@ ContratRepository contratRepository;
 		}
 	}
 	public float getChiffreAffaireEntreDeuxDates(Date startDate, Date endDate){
-		float difference_In_Time = endDate.getTime() - startDate.getTime();
+		float difference_In_Time = (float)(endDate.getTime() - startDate.getTime());
 		float difference_In_Days = (difference_In_Time / (1000 * 60 * 60 * 24)) % 365;
 		float difference_In_months =difference_In_Days/30;
         List<Contrat> contrats=contratRepository.findAll();
